API-794: Add withContexts HoC
