+++ src/Akeneo/Apps/back/Infrastructure/Symfony/Resources/public/js/react/react-controller.ts
@@ -45,8 +45,6 @@ abstract class ReactController extends BaseController {
       return;
     }
 
-    console.log('ok');
-
     unmoundReactElementRef();
   }
 }
